DatabaseManager: Fix schema loading.
The schema text file had an error in it, causing the estimates table to not
be created.

In addition, the database manager was not executing both CREATE TABLE
statements, because the sqflite execute method will only execute the first
statement in a sql string.

Fix the schema text file.
Enhance DatabaseManager.createTables to parse a multi-statement sql string
and execute each statement individually.