MainScreen: Bind app bar sorting action.
Now that the MainScreen owns the Scaffold, it also owns the AppBar
and the app bar's action menu items - such as the sort button.
The problem is, the Engagements that need sorted are in a separate screen
and/or child widget. This makes it challenging to bind the function
between the action button and the sortable list of Engagements.

Define a PopupMenuButtonHandler class to encapsulate an event handler function
for a popup menu button.

Add a PopupMenuButtonHandler property to the MainScreen's PopupMenu button.
Bind an intermediate function between the popup menu button and the
PopupMenuButtonHandler property.

Pass the PopupMenuButtonHandler to the child screens, such as the
ActiveEngagementList screen.

Expand ActiveEngagementList to accept a PopupMenuButtonHandler object,
and assign it a function during initState.

References #147, #148