EngagementScreen: Add non-nullable engagement and estimates.
Let the EngagementScreen receive a concrete engagement and list of
Estimates upon instantiation. Initialize the state with the list
of estimates.

Remove the named route, as we must perform async work (loading the
Estimates from the db) and pass the result to the EngagementScreen
constructor.

In ActiveEngagementListScreen, use a manual Navigator push instead
of the named route.

References #158.